# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Enxero Platform Backend API",
    "version": "1.0.0",
    "description": "Comprehensive API documentation for the Enxero Platform Backend, including Authentication, User Management, Company Management, Employee Management, Payroll Management, Leave Management, Forms Management, File Management, Notifications, Audit Logging, Integrations, and System Configuration.",
    "contact": {
      "name": "Enxero Support",
      "url": "https://www.enxero.com/support",
      "email": "support@enxero.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the JWT token in the format \"Bearer <token>\""
      }
    },
    "schemas": {
...
      "PayrollConfig": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "companyId": {
            "type": "string",
            "format": "uuid"
          },
          "payFrequency": {
            "type": "string",
            "enum": [
              "weekly",
              "biweekly",
              "monthly"
            ]
          },
          "taxSettings": {
            "type": "object"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PayrollPeriod": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "companyId": {
            "type": "string",
            "format": "uuid"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "closed",
              "processed"
            ]
          },
          "totalAmount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PayrollRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "periodId": {
            "type": "string",
            "format": "uuid"
          },
          "grossSalary": {
            "type": "number"
          },
          "totalDeductions": {
            "type": "number"
          },
          "netPay": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processed",
              "paid"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/payroll/config": {
      "get": {
        "summary": "Retrieve payroll configuration for the current company",
        "tags": [
          "Payroll"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Payroll configuration retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayrollConfig"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create payroll configuration for the company",
        "tags": [
          "Payroll"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayrollConfig"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payroll config created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayrollConfig"
                }
              }
            }
          }
        }
      }
    },
    "/payroll/config/{id}": {
      "put": {
        "summary": "Update payroll configuration",
        "tags": [
          "Payroll"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "PayrollConfig UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayrollConfig"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payroll config updated successfully"
          }
        }
      }
    },
    "/payroll/periods": {
      "get": {
        "summary": "Retrieve a list of payroll periods",
        "tags": [
          "Payroll"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Filter by status (open, closed, processed)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of payroll periods"
          }
        }
      },
      "post": {
        "summary": "Create a new payroll period",
        "tags": [
          "Payroll"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayrollPeriod"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payroll period created successfully"
          }
        }
      }
    },
    "/payroll/periods/{id}": {
      "get": {
        "summary": "Get payroll period by ID",
        "tags": [
          "Payroll"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "PayrollPeriod UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Payroll period details"
          }
        }
      }
    },
    "/payroll/periods/{id}/process": {
      "post": {
        "summary": "Process payroll for a period",
        "tags": [
          "Payroll"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "PayrollPeriod UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Payroll processed successfully"
          }
        }
      }
    },
    "/payroll/periods/{id}/approve": {
      "post": {
        "summary": "Approve payroll for a period",
        "tags": [
          "Payroll"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "PayrollPeriod UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Payroll approved successfully"
          }
        }
      }
    },
    "/payroll/records": {
      "get": {
        "summary": "Retrieve a list of payroll records for the company",
        "tags": [
          "Payroll"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "periodId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by payroll period"
          },
          {
            "in": "query",
            "name": "employeeId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by employee"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Filter by status (pending, processed, paid)"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of payroll records"
          }
        }
      },
      "post": {
        "summary": "Create a new payroll record",
        "tags": [
          "Payroll"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayrollRecord"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payroll record created successfully"
          }
        }
      }
    },
    "/payroll/records/{id}": {
      "get": {
        "summary": "Get payroll record by ID",
        "tags": [
          "Payroll"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "PayrollRecord UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Payroll record details"
          }
        }
      },
      "put": {
        "summary": "Update payroll record by ID",
        "tags": [
          "Payroll"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "PayrollRecord UUID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayrollRecord"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payroll record updated successfully"
          }
        }
      }
    },
    "/payroll/employee/{employeeId}/period/{periodId}": {
      "get": {
        "summary": "Get payroll details for an employee in a specific period",
        "tags": [
          "Payroll"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "employeeId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Employee UUID"
          },
          {
            "in": "path",
            "name": "periodId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "PayrollPeriod UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Payroll details for the employee and period"
          }
        }
      }
    },
...
  },
...
}